Swap in new table-drawing code
