+++ src/encoded/static/components/experiment.js
@@ -15,7 +15,6 @@ var pipeline = require('./pipeline');
 var reference = require('./reference');
 var biosample = require('./biosample');
 var software = require('./software');
-var sortTable = require('./sorttable');
 
 var Breadcrumbs = navbar.Breadcrumbs;
 var DbxrefList = dbxref.DbxrefList;
@@ -33,8 +32,6 @@ var PubReferenceList = reference.PubReferenceList;
 var ExperimentTable = dataset.ExperimentTable;
 var SingleTreatment = biosample.SingleTreatment;
 var SoftwareVersionList = software.SoftwareVersionList;
-var SortTablePanel = sortTable.SortTablePanel;
-var SortTable = sortTable.SortTable;
 
 
 var anisogenicValues = [
@@ -43,171 +40,6 @@ var anisogenicValues = [
     'anisogenic, sex-matched',
     'anisogenic'
 ];
-
-
-function humanFileSize(size) {
-    if (size === undefined) return undefined;
-    var i = Math.floor( Math.log(size) / Math.log(1024) );
-    return ( size / Math.pow(1024, i) ).toPrecision(3) * 1 + ' ' + ['B', 'kB', 'MB', 'GB', 'TB'][i];
-}
-
-
-// Configuration for process file table
-var rawTableConfig = {
-    title: 'Raw data',
-    columns: {
-        'accession': {
-            title: 'Accession',
-            display: function(item) {
-                return (
-                    <span>
-                        {item.title}<br />
-                        <a href={item.href} download={item.href.substr(item.href.lastIndexOf("/") + 1)} data-bypass="true"><i className="icon icon-download"></i> Download</a><br />
-                        {humanFileSize(item.file_size)}
-                    </span>
-                );
-            }
-        },
-        'file_type': {
-            title: 'File type'
-        },
-        'biological_replicates': {
-            title: function(list, config, meta) {
-                return (
-                    <span>{meta.anisogenic ? 'Anisogenic' : 'Biological'} replicate</span>
-                );
-            },
-            getValue: function(item) {
-                return item.biological_replicates ? item.biological_replicates.sort(function(a,b){ return a - b; }).join(', ') : '';
-            }
-        },
-        'technical_replicate_number': {
-            title: 'Technical replicate',
-            getValue: function(item) {
-                return item.replicate ? item.replicate.technical_replicate_number : null;
-            }
-        },
-        'read_length': {
-            title: 'Read length',
-            display: function(item) {
-                return <span>{item.read_length ? <span>{item.read_length + ' ' + item.read_length_units}</span> : null}</span>;
-            }
-        },
-        'run_type': {
-            title: 'Run type'
-        },
-        'paired_end': {
-            title: 'Paired end'
-        },
-        'assembly': {
-            title: 'Mapping assembly'
-        },
-        'title': {
-            title: 'Lab',
-            getValue: function(item) {
-                return item.lab && item.lab.title ? item.lab.title : null;
-            }
-        },
-        'date_created': {
-            title: 'Date added',
-            getValue: function(item) {
-                return moment.utc(item.date_created).format('YYYY-MM-DD');
-            },
-            objSorter: function(a, b) {
-                if (a.date_created && b.date_created) {
-                    return Date.parse(a.date_created) - Date.parse(b.date_created);
-                }
-                return a.date_created ? -1 : (b.date_created ? 1 : 0);
-            }
-        },
-        'status': {
-            title: 'Validation status',
-            display: function(item) {
-                return <div className="characterization-meta-data"><StatusLabel status="pending" /></div>;
-            },
-            hide: function(list, config, meta) {
-                return meta.encodevers !== '3';
-            },
-            sorter: false
-        }
-    }
-};
-
-
-// Configuration for process file table
-var processTableConfig = {
-    title: 'Processed data',
-    columns: {
-        'accession': {
-            title: 'Accession',
-            display: function(item) {
-                return (
-                    <span>
-                        {item.title}<br />
-                        <a href={item.href} download={item.href.substr(item.href.lastIndexOf("/") + 1)} data-bypass="true"><i className="icon icon-download"></i> Download</a><br />
-                        {humanFileSize(item.file_size)}
-                    </span>
-                );
-            }
-        },
-        'file_type': {
-            title: 'File type'
-        },
-        'output_type': {
-            title: 'Output type'
-        },
-        'biological_replicates': {
-            title: function(list, config, meta) {
-                return (
-                    <span>{meta.anisogenic ? 'Anisogenic' : 'Biological'} replicate</span>
-                );
-            },
-            getValue: function(item) {
-                return item.biological_replicates ? item.biological_replicates.sort(function(a,b){ return a - b; }).join(', ') : '';
-            }
-        },
-        'technical_replicate_number': {
-            title: 'Technical replicate',
-            getValue: function(item) {
-                return item.replicate ? item.replicate.technical_replicate_number : null;
-            }
-        },
-        'assembly': {
-            title: 'Mapping assembly'
-        },
-        'genome_annotation': {
-            title: 'Genome annotation'
-        },
-        'title': {
-            title: 'Lab',
-            getValue: function(item) {
-                return item.lab && item.lab.title ? item.lab.title : null;
-            }
-        },
-        'date_created': {
-            title: 'Date added',
-            getValue: function(item) {
-                return moment.utc(item.date_created).format('YYYY-MM-DD');
-            },
-            sorter: function(a, b) {
-                if (a.date_created && b.date_created) {
-                    return Date.parse(a.date_created) - Date.parse(b.date_created);
-                }
-                return a.date_created ? -1 : (b.date_created ? 1 : 0);
-            }
-        },
-        'status': {
-            title: 'Validation status',
-            display: function(item) {
-                return <div className="characterization-meta-data"><StatusLabel status="pending" /></div>;
-            },
-            hide: function(list, config, meta) {
-                return meta.encodevers !== '3';
-            },
-            sorter: false
-        }
-    }
-};
 
 
 var Panel = function (props) {
@@ -322,17 +154,6 @@ var Experiment = module.exports.Experiment = React.createClass({
         ];
 
         var experiments_url = '/search/?type=experiment&possible_controls.accession=' + context.accession;
-
-        // Extract three kinds of file arrays
-        var files = _(context.files).groupBy(function(file) {
-            if (file.output_category === 'raw data') {
-                return 'raw';
-            } else if (file.output_category === 'reference') {
-                return 'ref';
-            } else {
-                return 'proc';
-            }
-        });
 
         // XXX This makes no sense.
         //var control = context.possible_controls[0];
@@ -515,18 +336,8 @@ var Experiment = module.exports.Experiment = React.createClass({
 
                 {context.files.length ?
                     <div>
-                        <h3>Old Files linked to {context.accession}</h3>
-                        <FileTable items={context.files} encodevers={encodevers} anisogenic={anisogenic} />
-                    </div>
-                : null }
-
-                {context.files.length ?
-                    <div>
                         <h3>Files linked to {context.accession}</h3>
-                        <SortTablePanel>
-                            <SortTable list={files.raw} config={rawTableConfig} meta={{encodevers: encodevers, anisogenic: anisogenic}} />
-                            <SortTable list={files.proc} config={processTableConfig} meta={{encodevers: encodevers, anisogenic: anisogenic}} />
-                        </SortTablePanel>
+                        <FileTable items={context.files} encodevers={encodevers} anisogenic={anisogenic} />
                     </div>
                 : null }
